requirements/base.txt
Django==2.1.3
faust==1.3.3
kafka-python==1.4.4
gevent==1.3.7
redis==3.0.1
pymongo==3.7.2